BWEM: Minor changes

+++ BWAPI4J/src/main/java/bwem/util/Utils.java
@@ -13,7 +13,7 @@ public final class Utils {
     }
 
     public static int squaredNorm(int dx, int dy) {
-        return ((dx * dx) + (dy * dy));
+        return (int) ((dx * dx) + (dy * dy));
     }
 
     public static double norm(int dx, int dy) {
